OneDrive/Documents/python/DL/projects/HG_Detection/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=11.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
